variables types update

+++ Firmware/lib/Simulator/Simulator.cpp
@@ -1463,7 +1463,7 @@ void SimulatorClass::Main()
       while (PendingRecord.llTriggerTime != 0 && PendingRecord.llTriggerTime <= (long long)(GET_MICROS - (RaceHandler.llRaceStartTime - 10000))) //10ms advance added
       {
          //ESP_LOGD(__FILE__, "%lld Pending record S%d TriggerTime %lld | %lld", GET_MICROS, PendingRecord.iSensorNumber, RaceHandler.llRaceStartTime + PendingRecord.llTriggerTime, PendingRecord.llTriggerTime);
-         RaceHandler._QueuePush({PendingRecord.iSensorNumber, (RaceHandler.llRaceStartTime + PendingRecord.llTriggerTime), PendingRecord.iState});
+         RaceHandler._QueuePush({PendingRecord.iSensorNumber, ((long long)RaceHandler.llRaceStartTime + (long long)PendingRecord.llTriggerTime), PendingRecord.iState});
          //And increase pending record
          _iDataPos++;
          PROGMEM_readAnything(&SimulatorQueue[_iDataPos], PendingRecord);
